requirements.txt
fastapi
uvicorn
scapy
mvt
pyjwt
